// web-application/src/types-openapi/api/models/ListServiceResponse.ts
export interface ListServiceResponse {
    /**
     * 
     * @type {Date}
     * @memberof ListServiceResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ListServiceResponse
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    status?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListServiceResponse
     */
    pricePerHour?: number;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    category?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    unit?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListServiceResponse
     */
    stockQuantity?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListServiceResponse
     */
    note?: string | null;
}